+++ styles/tailwind.config.js
@@ -1,4 +1,4 @@
-const mystTheme = require('@myst-theme/styles');
+const mystTheme = require('./index.js');
 
 module.exports = {
   darkMode: 'class',
@@ -10,6 +10,8 @@ module.exports = {
     '../packages/frontmatter/src/**/*.{js,ts,jsx,tsx}',
     '../docs/stories/**/*.{js,ts,jsx,tsx,mdx}',
   ],
-  theme: mystTheme.themeExtensions,
+  theme: {
+    extend: mystTheme.themeExtensions,
+  },
   plugins: [require('@tailwindcss/typography')],
 };